Fix the SSZ objects dependency
Fix the dependency between `BeaconBlockHeader`, `BeaconBlockBody` and
`ProposerSlashing` by moving `BeaconBlockHeader` to
`eth2/beacon/types/block_headers.py`
